+++ imapserver/memory/selected.go
@@ -304,7 +304,7 @@ func (s *selected) Expunge(ctx context.Context, writer *imapserver.ExpungeWriter
 	}
 	s.session.account.mu.Unlock()
 	for _, uid := range removed {
-		if err := writer.WriteExpunge(ctx, uid); err != nil {
+		if err := writer.WriteExpunge(ctx, uid, nil); err != nil {
 			return err
 		}
 	}